requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
anyio=3.7.1=pyhd8ed1ab_0
argon2-cffi=23.1.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py38h0a891b7_2
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
babel=2.13.1=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=1.0=mkl
bleach=6.1.0=pyhd8ed1ab_0
brotli-python=1.0.9=py38hfa26641_7
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.7.22=hbcca054_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py38h5eee18b_3
charset-normalizer=3.3.1=pyhd8ed1ab_0
cudatoolkit=10.1.243=h8cb64d8_10
cycler=0.12.1=pyhd8ed1ab_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=pyhd8ed1ab_0
freetype=2.10.4=h0708190_1
giflib=5.2.1=h36c2ea0_2
gmp=6.2.1=h58526e2_0
gnutls=3.6.15=he1e5248_0
icu=67.1=he1b5a44_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_resources=6.1.0=pyhd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=5.5.5=py38hd0cf306_0
ipython=8.12.0=pyh41d4057_0
ipython_genutils=0.2.0=py_1
jedi=0.19.1=pyhd8ed1ab_0
jinja2=2.11.3=pyhd8ed1ab_2
jpeg=9e=h166bdaf_1
json5=0.9.14=pyhd8ed1ab_0
jsonschema=4.17.3=pyhd8ed1ab_0
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=5.5.0=py38h578d9bd_0
jupyter_server=1.13.5=pyhd3eb1b0_0
jupyterlab=3.3.2=pyhd3eb1b0_0
jupyterlab_server=2.10.3=pyhd3eb1b0_1
kiwisolver=1.4.4=py38h6a678d5_0
lame=3.100=h7f98852_1001
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libpng=1.6.39=h5eee18b_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h7f98852_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=8.0.1=hc9558a2_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=1.1.1=py38h497a2fe_3
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py38h5d868c9_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=0.8.4=pyh1a96a4e_1006
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpmath=1.3.0=pyhd8ed1ab_0
nbclassic=0.5.6=pyhb4ecaf3_1
nbconvert=5.6.1=py38h32f6830_1
nbformat=5.9.2=pyhd8ed1ab_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.8=pyhd8ed1ab_0
notebook-shim=0.2.3=pyhd8ed1ab_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
openh264=2.1.1=h4ff587b_0
openssl=3.0.11=h7f8727e_2
packaging=23.2=pyhd8ed1ab_0
pandoc=2.19.2=ha770c72_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py38h6a678d5_1
pip=23.3=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
platformdirs=3.11.0=pyhd8ed1ab_0
prometheus_client=0.18.0=pyhd8ed1ab_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.16.1=pyhd8ed1ab_0
pyparsing=3.1.1=pyhd8ed1ab_0
pyrsistent=0.18.0=py38heee7806_0
pysocks=1.7.1=pyha2e5f31_6
python=3.8.18=h955ad1f_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.18.1=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=2.1.0=py3.8_cpu_0
pytorch-mutex=1.0=cpu
pytz=2023.3.post1=pyhd8ed1ab_0
pyyaml=6.0=py38h0a891b7_4
pyzmq=19.0.2=py38ha71036d_2
readline=8.2=h5eee18b_0
requests=2.31.0=pyhd8ed1ab_0
send2trash=1.8.2=pyh41d4057_0
setuptools=68.0.0=py38h06a4308_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pyhd8ed1ab_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pyh04b8f61_3
tbb=2021.8.0=hdb19cb5_0
terminado=0.17.1=pyh41d4057_0
testpath=0.6.0=pyhd8ed1ab_0
tk=8.6.12=h1ccaba5_0
torchvision=0.16.0=py38_cpu
tornado=6.1=py38h0a891b7_3
traitlets=5.13.0=pyhd8ed1ab_0
typing-extensions=4.8.0=hd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
urllib3=2.0.7=pyhd8ed1ab_0
wcwidth=0.2.9=pyhd8ed1ab_0
webencodings=0.5.1=pyhd8ed1ab_2
websocket-client=1.6.4=pyhd8ed1ab_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7f98852_2
zeromq=4.3.4=h2531618_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
